+++ google-cloud-bigtable-bom/pom.xml
@@ -70,6 +70,11 @@
           <version>0.144.1-SNAPSHOT</version><!-- {x-version-update:google-cloud-bigtable-emulator:current} -->
         </dependency>
         <dependency>
+          <groupId>com.google.cloud</groupId>
+          <artifactId>google-cloud-bigtable-emulator-core</artifactId>
+          <version>0.144.1-SNAPSHOT</version><!-- {x-version-update:google-cloud-bigtable-emulator:current} -->
+        </dependency>
+        <dependency>
           <groupId>com.google.api.grpc</groupId>
           <artifactId>grpc-google-cloud-bigtable-admin-v2</artifactId>
           <version>2.7.1-SNAPSHOT</version><!-- {x-version-update:grpc-google-cloud-bigtable-admin-v2:current} -->
